third_party/blink/web_tests/external/wpt/css/css-borders/tentative/parsing/box-shadow-offset-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property box-shadow-offset value '0 0'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-offset value '10px 0'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-offset value '0 10px'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-offset value '10px 10px'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-offset value '10px 20px, 30px 40px'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property box-shadow-offset value 'calc(1em + 1px) calc(-1em + 1px)'
  assert_true: box-shadow-offset doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
